Overworld skeleton
Made skeleton for overworld which follows the player and then does nothing.
