+++ MonsterMash/Assets/Prefabs/Skeleton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7c76fa3a5a51f8942be4f507406c1d02
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
