+++ MonsterMash/Assets/Scripts/Player/EnemyOA.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d381248de8411d447b8b76a88cdaeeaa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
