+++ MonsterMash/Assets/Scripts/Player/OverworldAgent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 967e8ad0beff5ef41a41c6773663bd4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
