+++ MonsterMash/Assets/Sprites/enemies/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 057da2e9cf499df4ba28f731b7082e82
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
