+++ MonsterMash/Assets/Sprites/enemies/Animation/SkelAnimController.overrideController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7dd8dbcc1affbb745be1a4fb9f2507ab
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
