+++ MonsterMash/Assets/Sprites/enemies/Animation/SkelDownMove.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39f5dbdfd62167744ab3b21cd6f71280
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
