+++ MonsterMash/Assets/Sprites/enemies/Animation/SkelLeftMove.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a9c056d6012cd648ab085b884bbc4cc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
