+++ MonsterMash/Assets/Sprites/enemies/Animation/SkelRightMove.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 108e53843238d424aa42b33a3f6241e6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
