+++ MonsterMash/Assets/Sprites/enemies/Animation/SkelUpMove.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f020b63562483d841bea157765d67cc1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
